break up tests into component unit and service integration
